// apps/backend/src/shared/database/generated/client/models/StrategicInitiative.ts
export type StrategicInitiativeSelect<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetSelect<{
  id?: boolean
  seedKey?: boolean
  title?: boolean
  departmentId?: boolean
  createdQuarter?: boolean
  ownerName?: boolean
  year?: boolean
  comment?: boolean
  sourceTraceJson?: boolean
  version?: boolean
  createdAt?: boolean
  updatedAt?: boolean
  department?: boolean | Prisma.DepartmentDefaultArgs<ExtArgs>
}, ExtArgs["result"]["strategicInitiative"]>
// ...
export type StrategicInitiativeOmit<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetOmit<"id" | "seedKey" | "title" | "departmentId" | "createdQuarter" | "ownerName" | "year" | "comment" | "sourceTraceJson" | "version" | "createdAt" | "updatedAt", ExtArgs["result"]["strategicInitiative"]>
export type StrategicInitiativeInclude<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  department?: boolean | Prisma.DepartmentDefaultArgs<ExtArgs>
}
// ...
export type $StrategicInitiativePayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "StrategicInitiative"
  objects: {
    department: Prisma.$DepartmentPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    seedKey: string
    title: string
    departmentId: string
    createdQuarter: string
    ownerName: string
    year: string
    comment: string
    sourceTraceJson: string
    version: number
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["strategicInitiative"]>
  composites: {}
}

export type StrategicInitiativeGetPayload<S extends boolean | null | undefined | StrategicInitiativeDefaultArgs> = runtime.Types.Result.GetResult<Prisma.$StrategicInitiativePayload, S>
// ...
export type StrategicInitiativeDefaultArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the StrategicInitiative
   */
  select?: Prisma.StrategicInitiativeSelect<ExtArgs> | null
  /**
   * Omit specific fields from the StrategicInitiative
   */
  omit?: Prisma.StrategicInitiativeOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.StrategicInitiativeInclude<ExtArgs> | null
}